Correct key bindings for alt keyboard layouts
This fix uses the Unicode Utilities to check what letter is produced by
each key code in SDKeyBindingTranslator's initialize method. The values
are then stored in a NSMutableDictionary that is later used to check for
the proper key code for single character bindings. Only the subset of
key codes that can produce different characters for different keyboard
layouts are checked.